[BEAM-12339] Move unparse tests to different class to avoid 'pipeline not run' error.
